DefaultRegistrationEngine: Add support to match certificate usages for root CA
If Certificate Usage is CA contraint or trust anchor assertion the common name
in Identity is CA certificate.

Matching CN against server certificate is not going to match.

In X.509 certificate authentication DTLS has already verified the peers.

For X.509 it is sufficient to verify only the peer addresses that they match.

Signed-off-by: Vesa Jääskeläinen <dachaac@gmail.com>